Refs #37375 - Refactor Database class to avoid options hash
The options hash is a poor API because it's not immediately obvious what
the various parameters are. This refactors to be a very light wrapper
around Sequel.connect with optional support to migrate the old database.

The ContainerGatewayBackendTest is also modified to use an in-memory
database, which removes the need for any tear down.