Fixes #37375 - Refactor settings to connection_string
This prefers a single connection_string setting to be used to connect to
a database. This can be anything that Sequel supports. This simplifies
configuration and also allows for easy use of an in-memory database in
testing.

The default value for sqlite_db_path is dropped so that users explicitly
need to opt in to a migration if needed.

The example settings file now reflects this.